add and delete boarding

+++ Backend/app/controller/boardinghouse.controller.js
@@ -1,15 +1,20 @@
 const { BorardingHouse } = require("../models/index.model.js");
+
+const role = ["super-admin", "admin"];
+
 exports.create = async (req, res, next) => {
   const { name, address, rules } = req.body;
-  console.log("BorardingHouse Body:", req.body);
   try {
-    const document = await BorardingHouse.create({
-      name: name,
-      address: address,
-      rules: rules,
-      userId: req.userId,
-    });
-    res.json({ message: document, status: "success" });
+    if (role.includes(req.position.name)) {
+      const document = await BorardingHouse.create({
+        name: name,
+        address: address,
+        rules: rules,
+        userId: req.userId,
+      });
+      return res.json({ message: document, status: "success" });
+    }
+    return res.json({ message: "fail", status: "fail" });
   } catch (error) {
     console.log(error);
     res.json({ message: error, status: "faild" });
@@ -62,12 +67,16 @@ exports.updated = async (req, res, next) => {
 };
 exports.delete = async (req, res, next) => {
   try {
-    const document = await BorardingHouse.destroy({
-      where: {
-        _id: req.params.id,
-      },
-    });
-    res.json({ message: document, status: "success" });
+    if (role.includes(req.position.name)) {
+      const document = await BorardingHouse.destroy({
+        where: {
+          _id: req.params.id,
+        },
+      });
+      return res.json({ message: document, status: "success" });
+    } else {
+      return res.json({ message: "fail", status: "fail" });
+    }
   } catch (error) {
     console.log(error);
     res.json({ message: error, status: "faild" });
